==> Code/05-make-csvs.ipynb <==
pwd

import pandas as pd
import numpy as np
from numpy import load
from numpy import savetxt

### saving npy as csv - adhd

adhd_data = load('../../BC-sim/synth-data-01/tx_log_adhd1.npy')
adhd_data

print(type(adhd_data))

savetxt('tx_log_adhd1.csv', adhd_data, delimiter=',')

file = pd.read_csv('../../BC-sim/synth-data-01/tx_log_adhd1.csv')
print("\nOriginal file:")
print(file)

# adding header
headerList = ['Sphere 1', 'Sphere 2']
  
# converting data frame to csv
file.to_csv('tx_log_adhd1.csv', header=headerList, index=False)
  
# display modified csv file
file2 = pd.read_csv('tx_log_adhd1.csv')
print('\nModified file:')
print(file2)

df2 = file2
df2

data = load('../../BC-sim/synth-data-01/tx_log_adhd1.npy')
data
new = data[0:1]
print(new)
print(new.shape)

new_row = pd.DataFrame({'Sphere 1': new[0,0], 'Sphere 2': new[0,1]}, index=[0])
df_fix = pd.concat([new_row,df2.loc[:]]).reset_index(drop=True)
print (df_fix)

df_fix #YAY

df_fix.to_csv('../../BC-sim/tx_log_adhd.csv') #success -- all values and headers accounted forrrr

### saving npy as csv - td

td_data = load('../../BC-sim/synth-data-01/tx_log_td1.npy')
td_data

print(type(td_data))

savetxt('tx_log_td1.csv', td_data, delimiter=',')

file = pd.read_csv('../../BC-sim/synth-data-01/tx_log_td1.csv')
print("\nOriginal file:")
print(file)

# adding header
headerList = ['Sphere 1', 'Sphere 2']
  
# converting data frame to csv
file.to_csv('tx_log_td1.csv', header=headerList, index=False)
  
# display modified csv file
file2 = pd.read_csv('tx_log_td1.csv')
print('\nModified file:')
print(file2)

df2 = file2
df2

data = load('../../BC-sim/synth-data-01/tx_log_td1.npy')
data
new = data[0:1]
print(new)
print(new.shape)

new_row = pd.DataFrame({'Sphere 1': new[0,0], 'Sphere 2': new[0,1]}, index=[0])
df_fix2 = pd.concat([new_row,df2.loc[:]]).reset_index(drop=True)
print (df_fix2)

df_fix2 #YAY

df_fix2.to_csv('../../BC-sim/tx_log_td.csv') #success -- all values and headers accounted forrrr

### combine csvs

frames = [df_fix, df_fix2]
  
result = pd.concat(frames)
display(result)

result.to_csv('tx_log_adhd_and_td.csv')

stuff = pd.read_csv('../../BC-sim/synth-data-01/sim_all.csv')
stuff

stuff.drop('Unnamed: 0', inplace=True, axis=1)
stuff

stuff['Deformation'] = stuff['Sphere 2'] - stuff['Sphere 1']

stuff

stuff.to_csv('sim_all_def.csv')

stuff = pd.read_csv('../../BC-sim/synth-data-01/sim_all.csv')
stuff

stuff.drop('Unnamed: 0', inplace=True, axis=1)
stuff

stuff['S Deformation'] =  stuff['Sphere 2'] - stuff['Sphere 1'] # s or specific deformation 
stuff

stuff['S Deformation'].iloc[500:1000] = 0
stuff

zarray = np.array(stuff['Sphere 2'] - stuff['Sphere 1'])
zarray

newz = zarray[500:1000]
newz

newz.shape

arr = np.concatenate((newz, newz))
print(arr.shape)

stuff['Z Deformation'] = arr
stuff

stuff.to_csv('sim_all_def.csv')
