==> Code/06-compress-brain-data.ipynb <==
pwd

import os
import ants
import numpy as np
from tqdm import tqdm
import pandas as pd
from numpy import load

temp_patients = np.array([ants.image_read(f'../BC-sim-bigdata/synth-data-01/ADHD-sub-{s:04d}.nii.gz').numpy() for s in tqdm(range(500))])

temp_patients.shape

temp_controls = np.array([ants.image_read(f'../BC-sim-bigdata/synth-data-01/TD-sub-{s:04d}.nii.gz').numpy() for s in tqdm(range(500))])

temp_controls.shape

data = np.concatenate((temp_patients,temp_controls),axis=0)

data.shape

df = pd.read_csv('../../BC-sim/BC-sim-bigdata/synth-data-01/sim_all_def.csv')
df

controls = df['DX'].values==0
patients = df['DX'].values!=0

np.savez_compressed(file='../../BC-sim/BC-sim-bigdata/synth-data-01/sim-brain-array.npz',data=data,controls=controls,patients=patients)
